requirements.txt
Django==2.0.2
django-bootstrap-form==3.3
django-bootstrap4==0.0.6
django-user-accounts==2.0.3
html5lib==0.999999999
Pillow==5.0.0
requests==2.20
six==1.11.0
urllib3==1.22
django-extensions==2.0.0
django-organizations==1.0.0
django-crispy-forms==1.7.0
django-ajax-selects==1.7.0
django-machina==0.6.0
django-markdownx==2.0.23
django-taggit==0.22.2
python-decouple==3.1
django-imagekit==4.0.2
django-maintenance-mode==0.9.1
djangorestframework==3.8.2
django-slack-oauth==1.4.3
django-background-tasks==1.1.13
stripe==2.17.0
django-mathfilters==0.4.0
numpy==1.16
